Generate pdfs for all language variants
